[core] Extract SafeBotAuthSession implementation from AuthControl
