+++ mirai-core/src/commonMain/kotlin/network/auth/BotAuthSessionInternal.kt
